Added relations_id but stll Unknown access type record and other data related errors

+++ src/main/java/com/example/ecopr/controller/ProductController.java
@@ -20,7 +20,7 @@ public class ProductController {
     }
 
     @GetMapping("/search")
-    public List<Products> searchProductsByCategory(@RequestParam Long categoryId) {
+    public List<Products> searchProductsByCategory(@RequestParam int categoryId) { // Changed to int
         return productService.getProductsByCategory(categoryId);
     }
 }
